Make StaticEntity.client private All data is written/read with UTF-8 Centralize integer/float property parsing Use single-precision floating point arithmetic
git-svn-id: http://svn.code.sf.net/p/kolmafia/code@1399 b29ace70-8910-0410-8dcc-aa2fc6433167

+++ src/net/sourceforge/kolmafia/BuffRequestFrame.java
@@ -247,7 +247,7 @@ public class BuffRequestFrame extends KoLFrame
 
 					int price = offerings[i].getPrice();
 					int [] turns = offerings[i].getTurns();
-					String tooltip = price + " meat (" + FLOAT_FORMAT.format( (double)turns[0] / (double)price ) + " turns/meat)";
+					String tooltip = price + " meat (" + FLOAT_FORMAT.format( (float)turns[0] / (float)price ) + " turns/meat)";
 					checkboxes[i].setToolTipText( tooltip );
 
 					if ( turns.length > 1 )
